Replace EhCache for EhCache3 implementation
Replace EhCache for EhCache3 implementation, as the old EhCache is based
 on version 2, whereas we want to upgrade to EhCache version 3

#enhancement/remove-deprecated-code
