+++ modelling/src/test/java/org/axonframework/modelling/saga/repository/EhCacheSagaStoreTest.java
@@ -16,24 +16,34 @@
 
 package org.axonframework.modelling.saga.repository;
 
-import net.sf.ehcache.CacheManager;
 import org.axonframework.common.caching.Cache;
 import org.axonframework.common.caching.EhCacheAdapter;
+import org.ehcache.CacheManager;
+import org.ehcache.config.CacheConfiguration;
+import org.ehcache.config.builders.CacheConfigurationBuilder;
+import org.ehcache.config.builders.ResourcePoolsBuilder;
+import org.ehcache.core.Ehcache;
+import org.ehcache.core.EhcacheManager;
+import org.ehcache.core.config.DefaultConfiguration;
 import org.junit.jupiter.api.*;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Concrete implementation of the {@link CachingSagaStoreTest} using the {@link EhCacheAdapter}.
  *
  * @author Steven van Beelen
+ * @author Gerard Klijs
  */
-class EhCachingSagaStoreTest extends CachingSagaStoreTest {
+class EhCacheSagaStoreTest extends CachingSagaStoreTest {
 
     private CacheManager cacheManager;
-    private net.sf.ehcache.Cache ehCache;
+    private org.ehcache.core.Ehcache ehCache;
 
     @AfterEach
     void tearDown() {
-        cacheManager.shutdown();
+        cacheManager.close();
     }
 
     @Override
@@ -53,8 +63,18 @@ class EhCachingSagaStoreTest extends CachingSagaStoreTest {
     }
 
     private void buildEhCache() {
-        ehCache = new net.sf.ehcache.Cache("test", 100, false, false, 10, 10);
-        cacheManager = CacheManager.create();
-        cacheManager.addCache(ehCache);
+        Map<String, CacheConfiguration<?, ?>> caches = new HashMap<>();
+        DefaultConfiguration config = new DefaultConfiguration(caches, null);
+        cacheManager = new EhcacheManager(config);
+        cacheManager.init();
+        ehCache = (Ehcache) cacheManager
+                .createCache(
+                        "test",
+                        CacheConfigurationBuilder
+                                .newCacheConfigurationBuilder(
+                                        Object.class,
+                                        Object.class,
+                                        ResourcePoolsBuilder.heap(100L).build())
+                                .build());
     }
 }
